strict warnings fix

+++ test/custom_builds/c/safety_locks/src/main.c
@@ -27,7 +27,7 @@ void test_singlethreaded_no_conflicts(void) {
     ecs_set(world, e, Position, {10, 20});
     
     ecs_table_t *table = ecs_get_table(world, e);
-    int32_t pos_column = ecs_table_get_column_index(world, table, ecs_id(Position));
+    int16_t pos_column = (int16_t)ecs_table_get_column_index(world, table, ecs_id(Position));
     ecs_component_record_t *cr = flecs_components_get(world, ecs_id(Position));
     
     bool result;
@@ -83,7 +83,7 @@ void test_singlethreaded_with_conflicts(void) {
     ecs_set(world, e, Position, {10, 20});
     
     ecs_table_t *table = ecs_get_table(world, e);
-    int32_t pos_column = ecs_table_get_column_index(world, table, ecs_id(Position));
+    int16_t pos_column = (int16_t)ecs_table_get_column_index(world, table, ecs_id(Position));
     ecs_component_record_t *cr = flecs_components_get(world, ecs_id(Position));
     
     bool result;
@@ -142,7 +142,7 @@ void test_multithreaded_no_conflicts(void) {
     ecs_set(world, e, Position, {10, 20});
     
     ecs_table_t *table = ecs_get_table(world, e);
-    int32_t pos_column = ecs_table_get_column_index(world, table, ecs_id(Position));
+    int16_t pos_column = (int16_t)ecs_table_get_column_index(world, table, ecs_id(Position));
     ecs_component_record_t *cr = flecs_components_get(world, ecs_id(Position));
     
     bool result;
@@ -187,7 +187,7 @@ void test_multithreaded_with_conflicts(void) {
     ecs_set(world, e, Position, {10, 20});
     
     ecs_table_t *table = ecs_get_table(world, e);
-    int32_t pos_column = ecs_table_get_column_index(world, table, ecs_id(Position));
+    int16_t pos_column = (int16_t)ecs_table_get_column_index(world, table, ecs_id(Position));
     ecs_component_record_t *cr = flecs_components_get(world, ecs_id(Position));
     
     bool result;
@@ -229,7 +229,7 @@ void test_resizing(void) {
     ecs_set(world, e, Position, {10, 20});
     
     ecs_table_t *table = ecs_get_table(world, e);
-    int32_t pos_column = ecs_table_get_column_index(world, table, ecs_id(Position));
+    int16_t pos_column = (int16_t)ecs_table_get_column_index(world, table, ecs_id(Position));
     
     bool result;
